feat(scheduler): cross-tier failover and capacity_exhausted detection (#493)
When all models in the current tier are exhausted (429/quota), CSA now
falls back to models from adjacent tiers before reporting failure. Also
adds "capacity_exhausted"/"capacity exhausted" to gemini rate-limit
patterns to detect MODEL_CAPACITY_EXHAUSTED errors.

Changes:
- Add cross-tier fallback in decide_failover() when intra-tier options exhausted
- Add "capacity_exhausted" and "capacity exhausted" to gemini RATE_LIMIT_PATTERNS
- Extract failover tests to failover_tests.rs (keep failover.rs under 800 lines)
- Add 4 new tests: cross-tier fallback, all-tiers-exhausted, capacity patterns

Closes #493
