+++ crates/cli-sub-agent/src/debate_cmd_resolve.rs
@@ -98,7 +98,7 @@ pub(crate) fn resolve_debate_tool(
     if let Some(ref tier) = tier_name {
         if let Some(cfg) = project_config
             && let Some(resolution) =
-                crate::run_helpers::resolve_tool_from_tier(tier, cfg, parent_tool, whitelist)
+                crate::run_helpers::resolve_tool_from_tier(tier, cfg, parent_tool, whitelist, &[])
         {
             return Ok((
                 resolution.tool,
